- Implementado método delete MonsterCard

+++ src/main/java/com/lucas/yugiohcards/repository/MonsterCardRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MonsterCardRepository extends JpaRepository<MonsterCard, Integer> {
+public interface MonsterCardRepository extends JpaRepository<MonsterCard, Long> {
 }
